[dfs] remove nfs and fix warning in dfs v2.0.

+++ components/dfs/dfs_v2/filesystems/elmfat/dfs_elm.c
@@ -812,7 +812,7 @@ int dfs_elm_stat(struct dfs_dentry *dentry, struct stat *st)
     if (result == FR_OK)
     {
         /* convert to dfs stat structure */
-        st->st_dev = (dev_t)(dentry->mnt->dev_id);
+        st->st_dev = (dev_t)(size_t)(dentry->mnt->dev_id);
         st->st_ino = (ino_t)dfs_dentry_full_path_crc32(dentry);
 
         st->st_mode = S_IFREG | S_IRUSR | S_IRGRP | S_IROTH |
