+++ components/dfs/dfs_v2/filesystems/tmpfs/dfs_tmpfs.c
@@ -399,7 +399,7 @@ static int dfs_tmpfs_stat(struct dfs_dentry *dentry, struct stat *st)
     if (d_file == NULL)
         return -ENOENT;
 
-    st->st_dev = (dev_t)(dentry->mnt->dev_id);
+    st->st_dev = (dev_t)(size_t)(dentry->mnt->dev_id);
     st->st_ino = (ino_t)dfs_dentry_full_path_crc32(dentry);
 
     st->st_mode = S_IFREG | S_IRUSR | S_IRGRP | S_IROTH |
